feat: [sc-38] Queries with auto scoping of supplier_id (#13)
* fixed

* fix

+++ src/routes/admin/admin/admin.php
@@ -6,9 +6,9 @@ use App\Http\Controllers\Admin\AdminController;
 use Illuminate\Support\Facades\Route;
 
 Route::group(['prefix' => 'admin'], function () {
-    Route::get('/', [AdminController::class, 'index'])->name('admin.index');
-    Route::get('/{id}', [AdminController::class, 'show'])->name('admin.show');
-    Route::post('/', [AdminController::class, 'store'])->name('admin.store');
-    Route::put('{id}', [AdminController::class, 'update'])->name('admin.update');
-    Route::delete('{id}', [AdminController::class, 'destroy'])->name('admin.destroy');
+    Route::get('/', [AdminController::class, 'index'])->name('admin.admin.index');
+    Route::get('/{id}', [AdminController::class, 'show'])->name('admin.admin.show');
+    Route::post('/', [AdminController::class, 'store'])->name('admin.admin.store');
+    Route::put('{id}', [AdminController::class, 'update'])->name('admin.admin.update');
+    Route::delete('{id}', [AdminController::class, 'destroy'])->name('admin.admin.destroy');
 });
